SHA Hash Util
- added SHA-256
