+++ library/build.gradle
@@ -9,8 +9,8 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 29
-        versionCode 41
-        versionName "0.1.41"
+        versionCode 42
+        versionName "0.1.42"
     }
     buildTypes {
         release {
