--- library/src/main/java/com/noelchew/ncutils/crypto/SHA1Util.java
@@ -1,31 +1,0 @@
-package com.noelchew.ncutils.crypto;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class SHA1Util {
-    public static String getSha1Hash(final String s) {
-        final String SHA1 = "SHA-1";
-        try {
-            // Create SHA-1 Hash
-            MessageDigest digest = MessageDigest
-                    .getInstance(SHA1);
-            digest.update(s.getBytes());
-            byte messageDigest[] = digest.digest();
-
-            // Create Hex String
-            StringBuilder hexString = new StringBuilder();
-            for (byte aMessageDigest : messageDigest) {
-                String h = Integer.toHexString(0xFF & aMessageDigest);
-                while (h.length() < 2)
-                    h = "0" + h;
-                hexString.append(h);
-            }
-            return hexString.toString();
-
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-}
